Switches to a .net core build structure and system
VS2015 solution is still available, but packages are built with .net
core. Still supports net 4.5.1. NetStandard1.3 is needed for
NetworkTables, with an additional dependancy on NameResolution that I
want to get rid of at some point. NetworkTablesCore needs 1.5, but the
only thing that loses is UAP support, which you need to use the managed
version anyway for.